:package: Prepare for packaging
