+++ package.json
@@ -2,6 +2,20 @@
   "name": "@aibtcdev/tools",
   "version": "0.0.7",
   "private": false,
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "dependencies": {
     "@bitflowlabs/core-sdk": "^1.0.1",
     "@jingcash/core-sdk": "^0.1.4",
@@ -24,19 +38,6 @@
     "tsup": "^8.3.5",
     "typescript": "^5.0.0"
   },
-  "type": "module",
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "bin": {
-    "aibtc-wallet": "./src/stacks-wallet/index.ts",
-    "aibtc-bns": "./src/stacks-bns/index.ts",
-    "aibtc-nft": "./src/sip-009/index.ts"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
@@ -57,17 +58,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    },
-    "./dao": {
-      "types": "./dist/stacks-dao/index.d.ts",
-      "import": "./dist/stacks-dao/index.mjs",
-      "require": "./dist/stacks-dao/index.js"
-    }
   }
 }
