+++ src/index.ts
@@ -1,4 +1,16 @@
-// Re-export DAO SDK
-export * from "./stacks-dao";
+// src/index.ts
 
-// Future SDK exports will go here
+// Re-export everything from stacks-dao
+export {
+  DaoSDK,
+  Executor,
+  Treasury,
+  BankAccount,
+  Messaging,
+  Payments,
+} from "./stacks-dao";
+export * from "./stacks-dao/types";
+
+// Namespaced export for compatibility
+import * as daoNamespace from "./stacks-dao";
+export { daoNamespace as dao };
